+++ plugin/inner/config.go
@@ -1,6 +1,11 @@
 package pluginInternal
 
-import "github.com/hashicorp/go-plugin"
+import (
+	"os"
+	"strings"
+
+	"github.com/hashicorp/go-plugin"
+)
 
 const PluginName = "debugtalk"
 const RPCPluginName = PluginName + "_rpc"
@@ -14,4 +19,19 @@ var HandshakeConfig = plugin.HandshakeConfig{
 	ProtocolVersion:  1,
 	MagicCookieKey:   "HttpRunnerPlus",
 	MagicCookieValue: PluginName,
+}
+
+const hrpPluginTypeEnvName = "HRP_PLUGIN_TYPE"
+
+var hrpPluginType string
+
+func init() {
+	hrpPluginType = strings.ToLower(os.Getenv(hrpPluginTypeEnvName))
+	if hrpPluginType == "" {
+		hrpPluginType = "grpc" // default
+	}
+}
+
+func IsRPCPluginType() bool {
+	return hrpPluginType == "rpc"
 }